MGMT-15598: Bugfix - list all cluster manifests irrespective of type (#2265)
A recent change to Assisted Service changed the default behaviour of the
manifest list functionality. By default, the manifest list will only
return manifests that were supplied by the user and will exclude system
generated manifests.

This PR works in conjunction with a PR to the service (and corresponding
Python client) to supply an additional parameter
"include_system_generated" and set this to "true" so that
assisted-test-infra will fetch all manifests irrespective of type.