Add new log record for testcase id (#2350)
Current code in parallel may have duplicate threads id , when run in parallel
multiple proccesses created and threads per proccess with duplicate id.

Current patch adding new record to logger names test_id set to ""
In case we want to update the param we can set it in the begining of test.

When request fixture exists per cluster creation we take the request.node.name
(testcase name) as test_id.

I tested it in parallel run and it works but i think we need more tests.