Add commands  for system load during cluster installation.
This addon simulate load on nodes before and during installation.
The command load allows to:
- Load / increase CPU usage for nodes
- Load / increase RAM usage for nodes
- Load / limit network upload bandwidth for nodes
- Load / limit network download bandwidth for nodes

Each command get node invoker to run the command and support two actions:
- execute
- undo

The load simulator will allow us to install cluster similar to customers env
for example ,we can limit the traffic rate to 50kbps and the installation will be
slow and we will hit on timeout while downloading / configuring / finalzing.
In addition we will be able to simulate overload ram memory or busy disk with IO.

Example for options in a test:
	masters = cluster.nodes.get_masters()
        load_egress = load_system.LoadNetworkUpload(masters, "1.2mbps", "ens3")
        load_ingress = load_system.LoadNetworkDownload(masters, "1.2mbps", "ens3")
        load_egress.execute()
        load_ingress.execute()
        load_ingress.undo()
        load_egress.undo()