fix: update golang download link
We are currently hitting 403 errors while trying to pull the golang binary from the gcs golang bucket. The bucket we downloaded the golang binary from has proven to be unreliable and no official announcement has been identified around what is expected in terms of access to it, updating the url to download from the official download page from golang.

Signed-off-by: ehila <ehila@redhat.com>